Fix sourcemap urls.
